jwt/requirements.txt
pyzmq==19.0.0
datetime==4.3
pyjwt==1.7.1